Broke out the catalog reading code into a separate module data_read.py
